requirements.txt
virtualenv
asyncio
azure-storage-blob 
azure-eventhub-checkpointstoreblob-aio
azure-identity
azure-eventhub
configparser